All: rewrite from the ground up
The old code was a mess of experiments which were never cleaned up. New
code is the most straightforward way of doing things.

* Each type has its own function(s) to serialize and deserialize all of
which are exposed in the API.
* Where relevant a `deserialize*Alloc` function is added which allows
the deserializer to allocate memory
* Code is deduplicated by having a comptime parameter decide how futher
deserialization is handled (with or without an allocator)
* Compile errors are created as invalid {de,}serialization targets are
found; leading to more useful compile errors (with a semi useful trace)

Future work:
* Create a mechanism to override {de,}serialization